+++ src/cpp/humap_binding.cpp
@@ -41,7 +41,7 @@ namespace py = pybind11;
 PYBIND11_MODULE(_hierarchical_umap, m) {
 	
 	py::class_<humap::HierarchicalUMAP>(m, "HUMAP")
-		.def(py::init<string, py::array_t<double>, int, double, string, string, bool, bool>())
+		.def(py::init<string, py::array_t<float>, int, float, string, string, bool, bool>())
 		.def(py::init<>())
 		.def("fit", &humap::HierarchicalUMAP::fit)
 		.def("transform", &humap::HierarchicalUMAP::transform)
