+++ src/cpp/utils.h
@@ -62,7 +62,7 @@ struct SparseData
 {
   SparseData() {}
 
-  SparseData(vector<double> data_, vector<int> indices_): data(data_), indices(indices_) {} 
+  SparseData(vector<float> data_, vector<int> indices_): data(data_), indices(indices_) {} 
 
 
   /**
@@ -71,13 +71,13 @@ struct SparseData
   * @param index column index in the sparse matrix
   * @param value the non-zero value
   */
-  void push(int index, double value) {
+  void push(int index, float value) {
     data.push_back(value);
     indices.push_back(index);
   }
 
   // non-zero values
-  vector<double> data;
+  vector<float> data;
 
   // column indices
   vector<int> indices;
@@ -96,14 +96,14 @@ struct SparseData
  * @return Container with the resulting values
  */
 template<typename T>
-std::vector<double> linspace(T start_in, T end_in, int num_in)
+std::vector<float> linspace(T start_in, T end_in, int num_in)
 {
 
-  std::vector<double> linspaced;
+  std::vector<float> linspaced;
 
-  double start = static_cast<double>(start_in);
-  double end = static_cast<double>(end_in);
-  double num = static_cast<double>(num_in);
+  float start = static_cast<float>(start_in);
+  float end = static_cast<float>(end_in);
+  float num = static_cast<float>(num_in);
 
   if (num == 0) { return linspaced; }
   if (num == 1) 
@@ -112,7 +112,7 @@ std::vector<double> linspace(T start_in, T end_in, int num_in)
       return linspaced;
     }
 
-  double delta = (end - start) / (num - 1);
+  float delta = (end - start) / (num - 1);
 
   for(int i=0; i < num-1; ++i)
     {
@@ -227,22 +227,22 @@ std::vector<T> arrange_by_indices(const std::vector<T>& data, std::vector<int>& 
 // }
 
 // Converts an Eigen::Matrix to tuple format
-std::tuple<std::vector<int>, std::vector<int>, std::vector<double>> to_row_format(const Eigen::SparseMatrix<double, Eigen::RowMajor>& M);
+std::tuple<std::vector<int>, std::vector<int>, std::vector<float>> to_row_format(const Eigen::SparseMatrix<float, Eigen::RowMajor>& M);
 
 // Creates an Eigen::Matrix from containers
-Eigen::SparseMatrix<double, Eigen::RowMajor> create_sparse(vector<int>& rows, vector<int>& cols, vector<double>& vals, int size, int density);
+Eigen::SparseMatrix<float, Eigen::RowMajor> create_sparse(vector<int>& rows, vector<int>& cols, vector<float>& vals, int size, int density);
 
 // Creates an Eigen::Matrix from SparseData
-Eigen::SparseMatrix<double, Eigen::RowMajor> create_sparse(const vector<SparseData>& X, int size, int density);
+Eigen::SparseMatrix<float, Eigen::RowMajor> create_sparse(const vector<SparseData>& X, int size, int density);
 
 // Computes the squared distance of two points
-double rdist(const vector<double>& x, const vector<double>& y);
+float rdist(const vector<float>& x, const vector<float>& y);
 
 // Clip a gradient value
-double clip(double value);
+float clip(float value);
 
 // Computes the pairwise distance for a matrix of points
-vector<vector<double>> pairwise_distances(vector<vector<double>>& X);
+vector<vector<float>> pairwise_distances(vector<vector<float>>& X);
 
 // output verbosity
 void log(bool verbose, const string& message);
